add status to models
